perf: something

+++ src/models/model.rs
@@ -36,7 +36,7 @@ pub struct OutputParams {
     pub path: Option<PathBuf>,
     pub video_stream_ix: Option<usize>,
     pub audio_stream_ix: Option<usize>,
-    pub selected_range: Option<(f32, f32)>,
+    pub selected_range: Option<(f64, f64)>,
 }
 
 impl OutputParams {
